Visual studio build, runs

+++ moo/mooutils/moomarkup.h
@@ -147,7 +147,7 @@ MooMarkupDoc       *moo_markup_doc_new              (const char         *name);
 MooMarkupDoc       *moo_markup_parse_file           (const char         *filename,
                                                      GError            **error);
 MooMarkupDoc       *moo_markup_parse_memory         (const char         *buffer,
-                                                     int                 size,
+                                                     gssize              size,
                                                      GError            **error);
 gboolean            moo_markup_write_pretty         (MooMarkupDoc       *doc,
                                                      MooFileWriter      *writer,
